requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.27
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pkg_resources==0.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.4
typing_extensions==4.13.2
whitenoise==6.7.0
